backend/requirements.txt
django==4.2.17
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
psycopg2==2.9.3
djangorestframework-api-key==2.3.0
django-filter